Update index.md
Reworking this a bit.  I know it's a bit of 'marketing speak', but I thought it was important to start listing out some of the advantages of the LMS Ecosystem.  I think people need to know not just what the different components are, but what makes Lyrion different and what it allows people to do.