Update index.md
Added information about WiiM hardware players. They are LMS-compatible out-of-the-box and it's important the community support them.

Also included the idea of getting a Touch (or Radio) off eBay or using an RPi as a player which was already mentioned but not explained.

Please edit this however you see fit. The main point is to let people know there are viable hardware options out there as this can be a point of resistance for many.