Don’t do GPG signature checking on Fedora and derivatives
Fedora are aiming to enforce GPG signature checking on all RPM packages.  We don’t sign our packages, so that will fail.

https://fedoraproject.org/wiki/Changes/Enforcing_signature_checking_by_default

A better solution would, of course, be to actually sign the packages and distribute the keys, perhaps even via our own Yum repository.  Better still would be to get LMS included in Fedora, but for that we need to unbundle any remaining non-free firmware and, ideally, switch to upstream versions of all bundled CPAN modules.

Signed-off-by: Peter Oliver <git@mavit.org.uk>